% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcontrast.R
\name{cluster_latlon}
\alias{cluster_latlon}
\title{Cluster points based on latitude, longitude, and potentially other traits}
\usage{
cluster_latlon(
  latlon,
  mean_points_per_cluster = 10000,
  nclusters = NULL,
  split_traits = c("lat", "lon"),
  do_scale = TRUE
)
}
\arguments{
\item{latlon}{Data.frame with fields for taxon, lon, lat and perhaps other fields}

\item{mean_points_per_cluster}{How many points \emph{on average} to assign to each cluster -- some will have more or fewer}

\item{nclusters}{If specified, how many clusters to divide points into.}

\item{split_traits}{The columns to use for clustering.}

\item{do_scale}{If TRUE, normalizes each column before running}
}
\value{
The latlon data.frame with an additional column for cluster id and cluster_* columns having the raw mean of all points in that cluster for that trait (i.e., cluster_lat for the mean latitude).
}
\description{
To compare between regions, we must first have assigned points to regions. This function uses k-means to do this. By default you can use latitude and longitude, but you can optionally use other factors instead or in addition to these factors. For example, if you want the algorithm to attempt to not split species into separate clusters, you could make a column that converts the taxon to a number and use that number. If nclusters is specified, the function will split into that many clusters; if not, it will use whatever mean_points_per_cluster is set to in order to divide into clusters of approximately that size. If data are missing for any of the split_traits columns, that row will be deleted.
}
