% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcontrast.R
\name{compare_clusters}
\alias{compare_clusters}
\title{Compare spatial clusters}
\usage{
compare_clusters(
  phy,
  latlon,
  tiprates,
  rates = c("turnover", "net.div", "speciation", "extinct.frac", "extinction"),
  ndraws = 100,
  maxdepth = Inf,
  mincomparisons = 3
)
}
\arguments{
\item{phy}{Phylogeny in ape phylo format}

\item{latlon}{Data.frame with fields for taxon, lon, lat and perhaps other fields. Expects a cluster field giving the cluster id.}

\item{tiprates}{Data.frame with fields for taxon and rates at the tips}

\item{rates}{Character vector giving the names for the rate columns in tiprates}

\item{ndraws}{If at least one of the two clades being compared has more than one species, how many times to sample comparisons at random}

\item{maxdepth}{If finite, only compare clades that split from each other at that time point or younger}

\item{mincomparisons}{Only compare clusters if there are at least that many independent pairs per cluster}
}
\value{
A 2D list with comparisons FROM each cluster (rows) TO each cluster (columns). Each element is a data.frame with all the comparisons between each pair of regions
}
\description{
This is the main workhorse function: it compares pairs of clusters by subsampling a tree to only taxa present in at least one of those areas. Then it uses the sisters package to find all bifurcations where all taxa in one branch are in one area and all taxa in the other branch are in the other area (if one taxon is in both areas, it doesn't use this comparison). It takes the rate estimate for a taxon in the "target" cluster branch and subtracts the rate estimate for a taxon in the "focal" cluster branch (the columns and rows of the output matrix, respectively). If at least one of the clades being compared has more than one taxon, taxa are repeatedly sampled at random.
}
