% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcontrast.R
\name{compare_two_clades}
\alias{compare_two_clades}
\title{Compare rates between a pair of clades}
\usage{
compare_two_clades(
  focal.taxa,
  target.taxa,
  tiprates,
  phy_pruned,
  rates = c("turnover", "net.div", "speciation", "extinct.frac", "extinction"),
  ndraws = 100
)
}
\arguments{
\item{focal.taxa}{Character vector of taxon names in the focal region}

\item{target.taxa}{Character vector of taxon names in the target region}

\item{tiprates}{Data.frame with fields for taxon and rates at the tips}

\item{phy_pruned}{Phylogeny in ape phylo format pruned to just the taxa in the focal or target set}

\item{rates}{Character vector giving the names for the rate columns in tiprates}

\item{ndraws}{If at least one of the two clades being compared has more than one species, how many times to sample comparisons at random}
}
\value{
Data.frame comparing the pair of taxon sets
}
\description{
Get mean and sd of difference across all pairs in a sister comparison of target (state 1, right) minus focal (state 0, left)
}
