% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcontrast.R
\name{plot_map_signif}
\alias{plot_map_signif}
\title{Plot a map showing different rates with color}
\usage{
plot_map_signif(latlon, summaries, focal_rate = "net.div")
}
\arguments{
\item{latlon}{Data.frame with fields for taxon, lon, lat and perhaps other fields. Expects a cluster field giving the cluster id.}

\item{summaries}{Output list from summarize_cluster_results()}

\item{focal_rate}{Character string the names for the rate column to plot}
}
\value{
Nothing, though it uses print to plot the ggplot2 object
}
\description{
This computes a t-test comparing a region vs all other regions (this does not use the t-test results from summarize_cluster_results(), other than for getting the median estimate for the comparisons. It then shows the median, lower 95\% CI, and upper 95\% CI for this region vs all other regions.
}
