% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcontrast.R
\name{summarize_cluster_results}
\alias{summarize_cluster_results}
\title{Compute summaries of raw results}
\usage{
summarize_cluster_results(
  results,
  latlon,
  rates = c("turnover", "net.div", "speciation", "extinct.frac", "extinction")
)
}
\arguments{
\item{results}{Output list from compare_clusters()}

\item{latlon}{Data.frame with fields for taxon, lon, lat and perhaps other fields. Expects a cluster field giving the cluster id.}

\item{rates}{Character vector giving the names for the rate columns in tiprates}
}
\value{
List containing rate_difference, sign_test_p (p-values of the sign test), t_test_p (p-values of the t-test), t_test_estimate (the mean proportion), t_test_lower_ci (the lower 95\% CI for the mean proportion), t_test_upper_ci (the upper 95\% CI for the mean proportion), n_comparisons_list (the number of comparisons between each cluster pair)
}
\description{
The compare_clusters() function has a lot of detail: each sister group comparison between each pair of regions. This summarizes it further. It will do a t-test and a sign test on the direction of comparisons between each rate for each pair of regions (it does not do a bonferroni or other correction), get the CI for the t-test value comparison and point estimate, and more
}
